# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Verus Gateway API
  description: |
    A high-performance HTTP gateway for accessing files stored on the Verus blockchain.

    ## ⚠️ OUTDATED DOCUMENTATION
    **NOTE:** This OpenAPI specification is outdated and does not reflect the current API implementation.
    The actual API does not use `/v1/` or `/v2/` prefixes. All file endpoints use the format:
    - `/c/{chain}/file/{txid_or_filename}`
    - `/c/{chain}/meta/{txid}`
    - `/c/{chain}/file/{txid}` (HEAD request)

    This file needs to be updated to match the current implementation in `internal/http/server/server.go`.
    Please refer to the README.md for accurate API documentation.

    ## Features
    - Multi-chain support (VRSC, VRSCTEST, PBaaS chains)
    - File retrieval with optional encryption (viewing keys)
    - Built-in caching for high performance
    - Comprehensive metrics and health checks

    ## Authentication
    Currently, no authentication is required for public endpoints.
    Admin endpoints may require authentication in production deployments.

    ## Rate Limiting
    Rate limiting should be configured at the infrastructure level for production deployments.
  version: 1.0.0
  contact:
    name: Verus Gateway Support
    url: https://github.com/devdudeio/verus-gateway
  license:
    name: MIT
    url: https://github.com/devdudeio/verus-gateway/blob/master/LICENSE
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /admin/cache/stats:
    get:
      tags:
        - Admin
      summary: Get cache statistics
      description: Returns current cache usage statistics
      operationId: getCacheStats
      responses:
        '200':
          description: Cache statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  hits:
                    type: integer
                    example: 1234
                  misses:
                    type: integer
                    example: 56
                  size:
                    type: integer
                    description: Total cache size in bytes
                    example: 10485760
                  items:
                    type: integer
                    description: Number of cached items
                    example: 42
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/cache:
    delete:
      tags:
        - Admin
      summary: Clear entire cache
      description: Removes all entries from the cache
      operationId: clearCache
      responses:
        '200':
          description: Cache cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: cache cleared successfully
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/cache/{key}:
    delete:
      tags:
        - Admin
      summary: Delete cache entry
      description: Removes a specific entry from the cache
      operationId: deleteCacheEntry
      parameters:
        - name: key
          in: path
          required: true
          description: Cache key to delete
          schema:
            type: string
      responses:
        '200':
          description: Cache entry deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: cache entry {key} deleted successfully
        '500':
          $ref: '#/components/responses/InternalError'
components:
...
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Machine-readable error code
          example: file_not_found
        message:
          type: string
          description: Human-readable error message
          example: File not found on the blockchain
        request_id:
          type: string
          format: uuid
          description: Request ID for debugging
          example: 550e8400-e29b-41d4-a716-446655440000
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: internal_error
            message: An internal error occurred
            request_id: 550e8400-e29b-41d4-a716-446655440000